requirements/local.txt
# Local development dependencies go here
-r base.txt

# Discovered with 'pip freeze --local'
wheel==0.24.0

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.3.2

django-extensions==1.5.5